move lower hybrid frequency tests to test_frequencies.py
